requirements.txt
faster-whisper
webrtcvad
numpy
sounddevice
pyttsx3
edge-tts
requests
python-dotenv
sentence-transformers
deep-translator
whisper
torch
scikit-learn
scipy
tqdm
transformers
huggingface-hub
typing_extensions
Pillow
python-dateutil
faiss-cpu
pandas
scikit-image
wikipedia
edge-tts
fuzzywuzzy
python-Levenshtein
ultralytics
noisereduce
pyzbar
imutils
tensorflow
keras
easyocr
paddleocr